Added MCMC analysis
# **Interpretations**
Basyian Inference on how I understand it, is where we have a specific parameter $\theta$ where we define a prior assumption or guess for that parameter, where we would be defining a distribution for this assumption, suggesting how confident we are for this prior.

For instance, what we want to show is that there is relationship between ΔΔG and temperature. By the temperature-invariant ΔΔG hypothesis the relationship is assumed to be none, wherein ΔΔG and Temperature are independent form each other, by this we let the intercept to have a distribution centered around 0 with a normal distribution with a standard deviation of 1. However for the slope we defined a more strict normal distribution, where instead having a standard deviation of 1 we defined it to be 0.1, further supporting the temperature-invariant ΔΔG hypothesis.

Given this we would have a model that is modeled with the following:
$$△△G = \beta_0 + \beta_1\cdot T+𝛆$$
Further, based on our prior belief, when we scatter plot these values in a grid, we should see that the error follows a bell-like distribution, we used a distribution of student-t as it allows the model to be more robust to outliers and extreme values.

Now, given that we started on an assumption that there is no relationship, and then we introduce our data we have produced, based on the summary showed our best estimate for the slope be 0.02, with a confidence interval from 0.01 till 0.02. With an estimate error of 0, it indicates a very high confidence on the given estimate. Moreover, we can observe that the confidence interval does not cover 0, which is interesting because we started with saying the the slope $\beta_1$ is centered around 0, wherein we also even defined it to be very strict with a standard deviation 0.01 which is close to 0, however the model showed that it isn't centered around 0, thus suggesting a relationship between the two variables.

The MCMC method used was a built-in algorithm within brms, it is called No-U Turn Sampler (NUTS) where it is more efficient and robust, and moreover achieves that idea of MCMC where we define a markov chain and then starting with a guess you try to use that information whether or not it is a candidate to stay within the markov chain, thus giving a distribution based on how well it fits with the given data and prior belief.